Corregido error en la recepción de las líneas de pedido. Se debía a incosistencia de tipos entre BBDD y API, se ha cambiado el tipo de dato del modelo de la API a decimal, para ganar precisión en las operaciones

+++ Models/PedidoLin.cs
@@ -5,11 +5,11 @@ public class PedidoLin{
     public int? idLineaPedido {get; set;}
     public int? idPedido {get; set;}
     public int? idProducto {get; set;}
-    public double? precio {get; set;}
-    public double? descuento {get; set;}
+    public decimal? precio {get; set;}
+    public decimal? descuento {get; set;}
     public int? idTipoIVA {get; set;}
     public int? cantidad {get; set;}
-    public double? totalLinea {get; set;}
+    public decimal? totalLinea {get; set;}
     public bool activo {get; set;}
     
 
@@ -19,7 +19,7 @@ public class PedidoLin{
         activo = true;
     }
 
-    public PedidoLin(int _idLineaPedido, int _idPedido, int _idProducto, double _precio, double _descuento, int _tipoIVA)
+    public PedidoLin(int _idLineaPedido, int _idPedido, int _idProducto, decimal _precio, decimal _descuento, int _tipoIVA)
     {
         idLineaPedido = _idLineaPedido;
         idPedido = _idPedido;
